Progress from FA 2022

+++ cd_spec_viewer_web/cd_spec_viewer_web/cdspec/views.py
@@ -1,16 +1,33 @@
-import datetime
+"""
+This file defines views to be used by the CDSpec app;
+Django views are python functions which take http requests, and return http responses (e.g. HTML);
+To execute a view, we must call it via a URL;
+Therefore, the cdspec urls.py will make calls to the views.py class to access views of the cdspec folder;
+More information: https://www.w3schools.com/django/django_views.php
 
+Prepared by COSC 401 as part of the CDSpec Viewer project for Dr. Sherrer
+
+This file and all contributions herin are covered by the GPL 3.0 License 
+https://www.gnu.org/licenses/gpl-3.0.html
+"""
+#Python
+import datetime
+#Django
 from django.shortcuts import render, get_object_or_404
 from django.views import generic
 from django.http import HttpResponse, HttpResponseRedirect
 from django.urls import reverse
 from django.views.decorators.http import require_http_methods
+from django.contrib.auth.decorators import permission_required
 from django_datatables_view.base_datatable_view import BaseDatatableView
 from django.contrib import messages
-
-from cd_spec_viewer_web.cdspec.models import SpecRun
+from django.contrib.auth import get_user_model
+#Local Django
+from cdspec.models import SpecRun
 from .forms import CreateForm, EditForm
-from cd_spec_viewer_web.cdspec.util import handle_file_upload, Units, graph_format
+from cdspec.util import handle_file_upload, Units, graph_format
+
+from django.db.models import Q
 
 # Create your views here.
 
@@ -19,16 +36,32 @@ class IndexView(generic.ListView):
     template_name = "cdspec/index.html"
     context_object_name = 'latest_runs'
 
+    def get(self, request, *args, **kwargs):
+        if kwargs:
+           return render(request, 'cdspec/index.html', {'username' : kwargs['user']})
+        else:
+           return render(request, 'cdspec/index.html')
+
     def get_queryset(self):
-        return SpecRun.objects.order_by('-upload_date')[:10]
+        #if uploadedby=user argument is passed, filter the table
+        if self.kwargs:
+           user = get_user_model().objects.get(username=self.kwargs['user'])
+           return SpecRun.objects.filter(upload_user=user).order_by('-upload_date')[:10]
+        else:
+           return SpecRun.objects.order_by('-upload_date')[:10]
 
 #Edit view, allows the editing of existing objects
 def edit(request, pk):
+    user = request.user
+    if not user.has_perm('cdspec.can_edit'):
+       messages.info(request, "You do not have permission to edit this model")
+       return HttpResponseRedirect("/cdspec/" + str(pk)) 
+    
     #The post statement is the form submit handler. 
     if request.method == 'POST':
-        #We first recreate the form object using the request objects.
+        """First recreates the form object using the request objects"""
         form = EditForm(request.POST, instance=get_object_or_404(SpecRun, pk=pk))
-        #As long as the form is valid, we proceed to parsing.
+        """If form is valid, it proceed to parsing"""
         if form.is_valid():
             model = form.save()
             return HttpResponseRedirect(reverse('cdspec:detail', args=(model.id,)))
@@ -39,6 +72,11 @@ def edit(request, pk):
 
 #Create View, allows the creation of new objects
 def create(request):
+    user = request.user
+    if not user.has_perm('cdspec.can_upload'):
+       messages.info(request, "You do not have permission to upload")
+       return HttpResponseRedirect("/cdspec/") 
+
     #The post statement is the form submit handler. 
     if request.method == 'POST':
         #We first recreate the form object using the request objects.
@@ -46,42 +84,82 @@ def create(request):
         #As long as the form is valid, we proceed to parsing.
         if form.is_valid():
             #We parse the file into three dictionaries, header, data and indicies all within parsed_dictionary
-            parsed_dictionary = handle_file_upload(request.FILES['source_file'])  
-            #We then save the form but don't commit to db yet
-            model = form.save(commit=False)
-            #We add all the model's fields that are from the parsed dictionary
-            date_time_string = parsed_dictionary['header']['DATE'] + " " + parsed_dictionary['header']['TIME']
-            model.run_date = datetime.datetime.strptime(date_time_string, "%y/%m/%d %H:%M:%S")
-            model.data = parsed_dictionary['data']
-            model.data_points = parsed_dictionary['header']['NPOINTS']
-            #Setting the x/y header names
-            if "XUNITS" in parsed_dictionary['header']: 
-                model.x_units = parsed_dictionary['header']['XUNITS']
-            if "YUNITS" in parsed_dictionary['header']:
-                model.y_units = parsed_dictionary['header']['YUNITS']
-            if "Y2UNITS" in parsed_dictionary['header']:
-                model.y2_units = parsed_dictionary['header']['Y2UNITS']
-            if "Y3UNITS" in parsed_dictionary['header']:
-                model.y3_units = parsed_dictionary['header']['Y3UNITS']
-            #print(molar_ellipticity_calculation(parsed_dictionary['data'], model.pathlength, model.protein_concentration, model.number_of_amino_acids, model.degrees_index))
-            #Then save the model to the db, here we can return a different view, maybe redirect.
-            model.save()
-            return HttpResponseRedirect(reverse('cdspec:detail', args=(model.id,)))
+            try:
+               parsed_dictionary = handle_file_upload(request.FILES['source_file'])  
+               #We then save the form but don't commit to db yet
+               model = form.save(commit=False)
+               #We add all the model's fields that are from the parsed dictionary
+               date_time_string = parsed_dictionary['header']['DATE'] + " " + parsed_dictionary['header']['TIME']
+               model.run_date = datetime.datetime.strptime(date_time_string, "%y/%m/%d %H:%M:%S")
+               model.data = parsed_dictionary['data']
+               model.data_points = parsed_dictionary['header']['NPOINTS']
+               #Setting the x/y header names
+               if "XUNITS" in parsed_dictionary['header']: 
+                   model.x_units = parsed_dictionary['header']['XUNITS']
+               if "YUNITS" in parsed_dictionary['header']:
+                   model.y_units = parsed_dictionary['header']['YUNITS']
+               if "Y2UNITS" in parsed_dictionary['header']:
+                   model.y2_units = parsed_dictionary['header']['Y2UNITS']
+               if "Y3UNITS" in parsed_dictionary['header']:
+                   model.y3_units = parsed_dictionary['header']['Y3UNITS']
+               #print(molar_ellipticity_calculation(parsed_dictionary['data'], model.pathlength, model.protein_concentration, model.number_of_amino_acids, model.degrees_index))
+               model.upload_user = user
+               model.upload_user_string = user.username
+               #Then save the model to the db, here we can return a different view, maybe redirect.
+               model.save()
+               return HttpResponseRedirect(reverse('cdspec:detail', args=(model.id,)))
+            except:
+               messages.error(request, 'Unable to parse file, format error')
+               return render(request, 'cdspec/create.html', {'form' : form})
+
     else:
         form = CreateForm()
     return render(request, 'cdspec/create.html', {'form': form,})
 
 #Singular View w/ graph
 def detail(request, pk):
+    user = request.user
     model = get_object_or_404(SpecRun, pk=pk)
+    
+    #restricts viewing access from users with no permission
+    if not model.visible_public:
+       if (not model.visible_student and not user.has_perm('cdspec.can_view_all')) or (model.visible_student and not user.has_perm('cdspec.can_view_student')):
+          messages.info(request, "You do not have permission to access this spec model")
+          return HttpResponseRedirect('/cdspec/')
+
     #Send the model and all of the data points (presented as a list of x, list of y, list of y2, list of y3) to the corresponding template
-    return render(request, 'cdspec/detail.html', {'specrun': model, 'x': graph_format(model.data, 0), 'y': graph_format(model.data, 1), 'y2': graph_format(model.data, 2), 'y3': graph_format(model.data, 3), "pk": pk})
+    return render(request, 'cdspec/detail.html', {'specrun': model, 'x': graph_format(model.data, 0), 'y': graph_format(model.data, 1), 'y2': graph_format(model.data, 2),
+    'y3': (graph_format(model.data, 3) if model.y3_units is not None else None), "pk": pk})
 
 #Multi View
 def multi(request, pks):
+
+    user = request.user
+
+    if pks == "":
+       messages.info(request, "Select table rows to use the Multi-Graph function")
+       return HttpResponseRedirect('/cdspec/')
+
     proteins = []
     for pk in pks.split('/')[:-1]:
-        proteins.append(get_object_or_404(SpecRun, pk=pk))
+        obj = get_object_or_404(SpecRun, pk=pk)
+        
+        #view all spec runs
+        if user.has_perm('cdspec.can_view_all'):
+           proteins.append(obj)
+           
+        #view spec runs only if visible to student or public
+        elif user.has_perm('cdspec.can_view_student'):
+           if obj.visible_student or obj.visible_public:
+              proteins.append(obj)
+              
+        #only view spec runs visible to public
+        else:
+           if obj.visible_public:
+              proteins.append(obj)
+           else:
+              messages.info(request, "You do not have permission to access this spec model")
+              return HttpResponseRedirect('/cdspec/')
 
     #check if the models have the same units
     x_units = proteins[0].x_units
@@ -91,12 +169,13 @@ def multi(request, pks):
     for protein in proteins:
         if protein.x_units != x_units or protein.y_units != y_units or protein.y2_units != y2_units or protein.y3_units != y3_units:
             messages.info(request, 'Multi-graph failed: graphs have different axes')
-            return HttpResponseRedirect('/')
+            return HttpResponseRedirect('/cdspec/')
 
     #Send the models and all of the data points (presented as a list of x, list of y, list of y2, list of y3) to the corresponding template
     output_object = [];
     for protein in proteins:
-        output_object.append({'run_title' : protein.run_title, 'model' : protein, 'x' : graph_format(protein.data, 0), 'y' : graph_format(protein.data, 1), 'y2' : graph_format(protein.data, 2), 'y3' : graph_format(protein.data, 3)});
+        output_object.append({'run_title' : protein.run_title, 'model' : protein, 'x' : graph_format(protein.data, 0), 'y' : graph_format(protein.data, 1), 'y2' : graph_format(protein.data, 2),
+        'y3' : (graph_format(protein.data, 3) if protein.y3_units is not None else None)});
 
     return render(request, 'cdspec/multi.html', {'proteins': output_object, 'pks': pks, 'first': proteins[0]})
 
@@ -104,17 +183,45 @@ def multi(request, pks):
 class SpecRunJson(BaseDatatableView):
     model = SpecRun
     
+    #filter out models based on logged in user and model visibility
+    def get_initial_queryset(self):
+        user = self.request.user
+        q = SpecRun.objects
+
+        #filter out if looking for models from single upload user
+        if self.kwargs:
+           q = q.filter(upload_user_string=self.kwargs['user'])
+
+        #view all spec runs
+        if user.has_perm('cdspec.can_view_all'):
+           return q
+        #view spec runs only if visible to student or public
+        elif user.has_perm('cdspec.can_view_student'):
+           return q.filter(Q(visible_student=True)|Q(visible_public=True))
+        #only view spec runs visible to public
+        else:
+           return q.filter(visible_public=True)
+        
+    
 #Delete view
 @require_http_methods(["POST"])
-def delete(request, pk): 
+def delete(request, pk):
+    user = request.user 
+    
     # fetch the object related to passed id 
-    obj = get_object_or_404(SpecRun, id = pk) 
-  
+    obj = get_object_or_404(SpecRun, id = pk)
+    
+    #If user has no permission delete
+    if not user.has_perm('cdspec.can_delete'):
+       messages.info(request, "You do not have permission to delete this model")
+       return HttpResponseRedirect("/" + str(pk)) 
   
     if request.method =="POST": 
-        # delete object 
+    
+        #delete object
+        obj.source_file.delete() 
         obj.delete() 
-        # after deleting redirect to  
-        # home page 
-        return HttpResponseRedirect("/") 
+        
+        # after deleting redirect to home page 
+        return HttpResponseRedirect("/cdspec/")
